+++ tests/loadtestdata.py
@@ -1,34 +1,27 @@
 import unittest
 
 from pandas import DataFrame
-import diive
 
-# from diive.core.io.filereader import ReadFileType
-# from diive.core.io.files import save_as_pickle, load_pickle
+from diive.core.io.filereader import ReadFileType
+from diive.core.io.files import load_pickle
 
-# import diive.core.io.files as files
-# from diive.core.io.filereader import ReadFileType
+def loadtestdata():
+    loaddatafile = ReadFileType(filetype='DIIVE_CSV_30MIN',
+                                filepath='../diive/configs/exampledata/exampledata_CH-DAV_FP2022.5_2022.07_ID20230206154316_30MIN.diive.csv',
+                                data_nrows=None)
+    data_df, metadata_df = loaddatafile._readfile()
+    return data_df, metadata_df
 
-
-class TestReadFileType(unittest.TestCase):
-
-    def load_testdata(self):
-        loaddatafile = ReadFileType(filetype='REDDYPROC_30MIN',
-                                    filepath='testdata/testdata_CH-DAV_FP2021.2_2016-2020_ID20220324003457_30MIN_SUBSET.csv',
-                                    data_nrows=None)
-        data_df, metadata_df = loaddatafile._readfile()
-        return data_df, metadata_df
+class TestLoadData(unittest.TestCase):
 
     def test_readfiletype(self):
-        data_df, metadata_df = self.load_testdata()
+        data_df, metadata_df = loadtestdata()
         self.assertEqual(type(data_df), DataFrame)
+        return data_df, metadata_df
 
-    def test_pickle(self):
-        data_df, metadata_df = self.load_testdata()
-        filepath = save_as_pickle(outpath='',
-                                  filename='testdata/testdata_CH-DAV_FP2021.2_2016-2020_ID20220324003457_30MIN_SUBSET.csv',
-                                  data=data_df)
-        load_pickle(filepath=filepath)
+    def test_load_pickle(self):
+        data_df = load_pickle(filepath='../diive/configs/exampledata/exampledata_CH-DAV_FP2022.5_2022_ID20230206154316_30MIN.diive.csv.pickle')
+        self.assertEqual(type(data_df), DataFrame)
 
 
 if __name__ == '__main__':